#17: reorganized actions
